pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2",
    "django-browser-reload>=1.18.0",
    "django-cors-headers>=4.7.0",
    "django-redis>=5.4.0",
    "django-rest-framework>=0.1.0",
    "django-tailwind>=4.0.1",
    "djangorestframework>=3.16.0",
    "faker>=37.1.0",
    "ipython>=9.2.0",
    "locust>=2.36.2",
    "pillow>=11.2.1",
]
